// src/pages/student/studentMentorTaType.d.ts
export type AssigneeStudentsAll = AssigneeStudent[]

export interface AssigneeStudent {
  personnelId: number
  moduleId: number
  personnelRole: string
  notes: any
  isActive: boolean
  id: number
  studentId: number
  isDeleted: boolean
  createdAt: string
  updatedAt: any
  deletedAt: any
  student: Student
  mentorsAndTAs: MentorsAndTa[]
  surveys: Survey[]
  enrollment: Enrollment
  module: Module
}

export interface Student {
  username: string
  idnumber: string
  firstName: string
  family: string
  email: string
  phone: string
  mobile: string
  institution: string
  department: string
  address: string
  city: string
  country: string
  lang: string
  timezone: string
  calendarType: string
  id: number
  picture?: Picture
  registrationForm: RegistrationForm
  statusForm: StatusForm
  careerPathway: CareerPathway2
}
// ...
export interface MentorsAndTa {
  assignmentId: number
  isActive: boolean
  studentId: number
  personnel: Personnel
  personnelRole: string
}

export interface Personnel {
  username: string
  idnumber: string
  firstName: string
  family: string
  email: string
  phone: string
  mobile: string
  institution: string
  department: string
  address: string
  city: string
  country: string
  lang: string
  timezone: string
  calendarType: string
  id: number
}

export interface Survey {
  isChecked: boolean
  comment: string
  studentContribution: string
  sessionDate: string
  studentPresent: string
  sessionProblem: string
  studentTask: string
  studentId: number
  moduleId: number
  commenterId: number
  isDeleted: boolean
  id: number
  createdAt: string
  updatedAt: any
  deleteTime: any
  commenterRole: string
}

export interface Enrollment {
  assessment: Assessment
  assignedTA: AssignedTa
  assignedMentor: any
}
// ...
export interface Module {
  instructorCount: number
  studentCount: number
  mentorCount: number
  teachingAssistantCount: number
  name: string
  description: any
  numberOfHours: number
  moduleType: string
  subType: string
  isActive: boolean
  isImported: any
  teachingStatus: string
  levelName: any
  nonLmsInstructors: any
  startDate: string
  endDate: string
  weblinkFeedbackForm: any
  weblinkFinalProject: any
  deadlineFinalProject: any
  weblinkLmsCourse: any
  id: number
  createdAt: string
  updatedAt: string
  category: Category
  careerPathway: CareerPathway3
  instructors: Instructor[]
}